chore: Event 엔티티의 userRatingSum 타입을 Long으로 변경

+++ src/main/java/com/quickticket/quickticket/domain/event/entity/Event.java
@@ -52,7 +52,7 @@ public class Event {
 
     @NotNull
     @Column(nullable = false)
-    private BigInteger userRatingSum;
+    private Long userRatingSum;
 
     @NotNull
     @Column(nullable = false)
